pyproject.toml
[tool.poetry]
name = "034-indexscraper"
version = "0.1.0"
description = ""
authors = ["nsamardzic13 <nikola.samardzic1997@gmail.com>"]
readme = "README.md"
packages = [{include = "034_indexscraper"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
pandas = "^2.2.0"
arrow = "^1.3.0"
scrapy = "^2.11.0"
boto3 = "^1.34.33"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
